refactor(messaging): split config status comparisons
